Trainer class almost complete.
Error handling, save-restore models, summaries and logging added to
training loop. Summaries defined in model are ignored but must be
carefully included in such a way that the user sees no change.

Trainer config is now build automatically except for the optimizer
field. This makes handling default config much easier.

Get summary op method added to model class.

Removed most of tensorflow names scopes for suspicion of making the
process somewhat slower. A way of making the simplified graph of the
model with the correct inputs must be deviced.

Distributed training is still missing. But not hurries.