Several fixes and small features.
1. Added a stop at iteration feature for the tensorflow trainer. This
option can now be specified at the feed portion of the train
configurations. Model training will stop at the specified training step.

2. Model objects now have an attribute called keep prob which handles
dropout probabilities. Such attribute should not be accesible to the
user. The predict method of Model instances handles the setting of this
variable to a user defined value (configurable in the regularization
section of the train config file) only when training.

3. Fixed bug that appeared when model summaries where included but no
user defined model summaries exists.

4. Regularization losses now appear in tensorboard if desired (must
specify in the summaries section of the train config file).