Proposed fix for object import error.
Object loading by axon suposes there is a single extension of an
abstract class in the corresponding file to be imported. Hence to load
it without explicitely giving the name of the desired object we look for
objects of the adequate class. This was badly implemented in the case of
multiple possible interfaces (abstract classes). The error is now
beleived to be fixed.